User Data Orginization && Queries
The user data is now being mapped and turned into a workable object for the client to consume.
Because of the nature of deeply nested folders a recursive approach was taken to manage the orginization of
nested objects. Another reason the workload is put on the server.... I better implement a great caching mechanism.

Notes and folders queries are now predefined and ready for testing. Not all queries will be used. This is a demonstration of knowledge.